@@ -1,112 +1,112 @@
-import React, {Component} from 'react';
+import React, { useState, useEffect } from 'react';
+import { useHistory } from 'react-router-dom';
+import { observer } from "mobx-react";
 import store from '../../stores/store';
-import {observer} from "mobx-react";
 import TopNav from "../../common/TopNav/TopNav";
 import FilterBar from "../../common/FilterBar/FilterBar";
 import JobsList from "../JobsList/JobsList";
 import Hero from "../../common/Hero";
-import {toJS} from "mobx";
 
-class JobsContainer extends Component {
-  state = {
-    confetti: false,
-    loadingResult: null
-  };
+const JobsContainer = (props) => {
+  const [confetti, setConfetti] = useState(false);
+  const [loadingResult, setLoadingResult] = useState(null);
+  const history = useHistory();
 
-  componentDidMount() {
-    const tech = this.props.match.params.tech;
-    const city = this.props.match.params.city;
-    const slug = this.props.match.params.id;
+  useEffect(() => {
+    const { tech, city, slug } = props.match.params;
 
     store.getCities();
     store.getHeroImage();
     store.currentJob = null;
-
     store.getTechnologies();
 
     if (slug) {
-      this.setState({loadingResult: store.showJobFromUrl(decodeURIComponent(slug))});
+      setLoadingResult(store.showJobFromUrl(decodeURIComponent(slug)));
       store.getJobsFromUrl();
-
     } else if (tech && city) {
-      let selectedTech;
-      if (tech) {
-        if (tech === 'all') {
-          store.allTechnologies = true;
-        } else {
-          store.allTechnologies = false;
-          selectedTech = tech.split(',');
-        }
-        store.filtering = true;
-      }
-
-      let remoteLevel;
-      if (city && city === 'remote') {
-        remoteLevel = true
-      }
-
-      store.getJobsFromUrlFiltered(selectedTech, remoteLevel);
-
+      handleTechAndCity(tech, city);
     } else if (tech && !city) {
-      let selectedTech;
-      if (tech === 'all') {
-        store.allTechnologies = true;
-      } else {
-        store.allTechnologies = false;
-        selectedTech = tech.split(',');
-      }
-      store.filtering = true;
-      store.getJobsFromUrlTechFiltered(selectedTech);
-
+      handleTechOnly(tech);
     } else {
       store.getAllJobs(10, 0);
     }
-  }
 
-  componentDidUpdate() {
-    window.onpopstate = (e) => {
+    window.onpopstate = () => {
       store.resetCurrentJob();
       store.mobileBar = null;
-    }
-  }
+    };
 
-  cleanCurrentJob() {
+    return () => {
+      // Cleanup event listeners
+      window.onpopstate = null;
+    };
+  }, []);
+
+  const handleTechAndCity = (tech, city) => {
+    let selectedTech;
+    if (tech === 'all') {
+      store.allTechnologies = true;
+    } else {
+      store.allTechnologies = false;
+      selectedTech = tech.split(',');
+    }
+    store.filtering = true;
+
+    let remoteLevel;
+    if (city === 'remote') {
+      remoteLevel = true;
+    }
+
+    store.getJobsFromUrlFiltered(selectedTech, remoteLevel);
+  };
+
+  const handleTechOnly = (tech) => {
+    let selectedTech;
+    if (tech === 'all') {
+      store.allTechnologies = true;
+    } else {
+      store.allTechnologies = false;
+      selectedTech = tech.split(',');
+    }
+    store.filtering = true;
+    store.getJobsFromUrlTechFiltered(selectedTech);
+  };
+
+  const cleanCurrentJob = () => {
     store.currentJobId = null;
     store.currentJob = null;
-  }
+  };
 
-  render() {
-    const {jobs, heroImage} = store;
-    const description = (
-      <>
-        Dev-specific filter job search.
-        <br/>
-        <br/>
-        Salary estimates.
-        <br/>
-        <br/>
-        Structured job ad design.
-      </>
-    );
+  const sideNavToggleHandler = () => {
+    store.mobileBar = !store.mobileBar;
+  };
 
-    return (
-      <>
-        <TopNav
-          onHamburgerClick={this.sideNavToggleHandler}
-        />
-        {!store.currentJob && heroImage &&
+  const description = (
+    <>
+      Dev-specific filter job search.
+      <br/>
+      <br/>
+      Salary estimates.
+      <br/>
+      <br/>
+      Structured job ad design.
+    </>
+  );
+
+  return (
+    <>
+      <TopNav onHamburgerClick={sideNavToggleHandler} />
+      {!store.currentJob && store.heroImage && (
         <Hero
           heading='IT jobs with salaries in&nbsp;Germany.'
           description={description}
-          heroImage={heroImage}
+          heroImage={store.heroImage}
         />
-        }
-        <FilterBar/>
-        <JobsList jobs={jobs}/>
-      </>
-    )
-  }
-}
+      )}
+      <FilterBar />
+      <JobsList jobs={store.jobs} />
+    </>
+  );
+};
 
-JobsContainer = observer(JobsContainer);
-export default JobsContainer;
+export default observer(JobsContainer);
