@@ -1,20 +1,19 @@
-import React from "react";
-import {NavLink} from "react-router-dom";
-import {ReactComponent as LogoDE} from 'assets/img/logo.svg';
-import {observer} from "mobx-react";
+import React, { useState, useContext } from "react";
+import { NavLink } from "react-router-dom";
+import { ReactComponent as LogoDE } from 'assets/img/logo.svg';
+import { observer } from "mobx-react";
 import style from './style.module.scss';
 import Button from '../Button/Button';
 import Hamburger from '../Hamburger/Hamburger';
 import SideNav from '../SideNav';
 import ThemeSwitcher from "../ThemeSwitcher";
-import {ThemeContext} from "../../themeContext";
+import { ThemeContext } from "../../themeContext";
 import MailchimpSubscribe from "react-mailchimp-subscribe";
-import {isMobile, isTablet} from "react-device-detect";
-
+import { isMobile } from "react-device-detect";
 
 const url = "https://jobsforit.us21.list-manage.com/subscribe/post?u=9bb02c9224c59d5a48a7cac52&id=9abe4768b0&f_id=00bbc1e1f0";
 
-const CustomForm = ({status, message, onValidated}) => {
+const CustomForm = ({ status, message, onValidated }) => {
   let email;
 
   const submit = () =>
@@ -85,47 +84,37 @@ const CustomForm = ({status, message, onValidated}) => {
   );
 };
 
-// const SubscribeForm = () =>
+const TopNav = observer(() => {
+  const [showSideNav, setShowSideNav] = useState(false);
+  const themeContext = useContext(ThemeContext);
 
-class TopNav extends React.Component {
-
-  state = {
-    showSideNav: false
+  const sideNavToggleHandler = () => {
+    setShowSideNav(prevState => !prevState);
   }
 
-
-  sideNavClosedHandler = () => {
-    this.setState({showSideNav: false});
+  const sideNavClosedHandler = () => {
+    setShowSideNav(false);
   }
 
-  sideNavToggleHandler = () => {
-    this.setState((prevState) => {
-      return {showSideNav: !prevState.showSideNav}
-    });
+  const classes = [style.topNav];
+  if (themeContext.theme === 'dark') {
+    classes.push(style.topNav_dark);
+  } else {
+    classes.push(style.topNav_light);
   }
 
-  render() {
-    const themeContext = this.context;
+  return (
+    <>
+      <div className={classes.join(' ')}>
+        <div className={style.topNav_wrapper}>
+          <div className={style.topNav_group}>
+            <a href='/' className={style.topNav_logo}>
+              <LogoDE />
+            </a>
+          </div>
 
-    const classes = [style.topNav];
-
-    if (themeContext.theme === 'dark') {
-      classes.push(style.topNav_dark);
-    } else {
-      classes.push(style.topNav_light);
-    }
-
-    return (
-      <>
-        <div className={classes.join(' ')}>
-          <div className={style.topNav_wrapper}>
+          {!isMobile && (
             <div className={style.topNav_group}>
-              <a href='/' className={style.topNav_logo}>
-                <LogoDE/>
-              </a>
-            </div>
-
-            { !isMobile && <div className={style.topNav_group}>
               <p style={{
                 marginRight: '1em',
                 fontSize: "1.5em",
@@ -135,7 +124,7 @@ class TopNav extends React.Component {
               }}>Get latest jobs</p>
               <MailchimpSubscribe
                 url={url}
-                render={({subscribe, status, message}) => (
+                render={({ subscribe, status, message }) => (
                   <CustomForm
                     status={status}
                     message={message}
@@ -143,50 +132,36 @@ class TopNav extends React.Component {
                   />
                 )}
               />
-            </div>}
-            <div className={style.topNav_group}>
-              <ThemeSwitcher/>
-              <ul className={style.topNav_menu}>
-                {/*<li className={style.topNav_menu_item}>*/}
-                {/*  <NavLink exact activeClassName={style.topNav_menu_item_activeLink} to={'/'}>Jobs</NavLink>*/}
-                {/*</li>*/}
-                {/*<li className={style.topNav_menu_item}>*/}
-                {/*<NavLink exact activeClassName={style.topNav_menu_item_activeLink} to={'/tutorials'}>DevTube</NavLink>*/}
-                {/*</li>*/}
-                <li className={style.topNav_menu_item}>
-                  <NavLink exact activeClassName={style.topNav_menu_item_activeLink}
-                           to={'/pricing'}>Pricing</NavLink>
-                </li>
-                <li className={style.topNav_menu_item}>
-                  <NavLink exact activeClassName={style.topNav_menu_item_activeLink}
-                           to={'/statistics'}>Statistics</NavLink>
-                </li>
-                {/*<li className={style.topNav_menu_item}>*/}
-                {/*  <NavLink exact activeClassName={style.topNav_menu_item_activeLink} to={'/token'}>Token</NavLink>*/}
-                {/*</li>*/}
-                <li className={style.topNav_menu_item}>
-                  {/*<NavLink exact activeClassName={style.topNav_menu_item_activeLink} to={'/brand-room'}>Brand Room</NavLink>*/}
-                </li>
-                <li className={style.topNav_menu_item}>
-                  <Button href={'/pricing'} classes={[style.topNav_button]}>Post a Job</Button>
-                </li>
-              </ul>
-              <Hamburger
-                onClick={this.sideNavToggleHandler}
-              />
             </div>
+          )}
+
+          <div className={style.topNav_group}>
+            <ThemeSwitcher />
+            <ul className={style.topNav_menu}>
+              <li className={style.topNav_menu_item}>
+                <NavLink exact activeClassName={style.topNav_menu_item_activeLink}
+                         to={'/pricing'}>Pricing</NavLink>
+              </li>
+              <li className={style.topNav_menu_item}>
+                <NavLink exact activeClassName={style.topNav_menu_item_activeLink}
+                         to={'/statistics'}>Statistics</NavLink>
+              </li>
+              <li className={style.topNav_menu_item}>
+                <Button href={'/pricing'} classes={[style.topNav_button]}>Post a Job</Button>
+              </li>
+            </ul>
+            <Hamburger
+              onClick={sideNavToggleHandler}
+            />
           </div>
         </div>
-        <SideNav
-          open={this.state.showSideNav}
-          onCloseClick={this.sideNavClosedHandler}
-        />
-      </>
-    )
-  }
-}
+      </div>
+      <SideNav
+        open={showSideNav}
+        onCloseClick={sideNavClosedHandler}
+      />
+    </>
+  );
+});
 
-TopNav.contextType = ThemeContext;
-
-TopNav = observer(TopNav);
 export default TopNav;
